Don't update bulkop records on migration requests
Fixes #15539
Fixes #15555

+++ src/services/Elements.php
@@ -17,6 +17,9 @@ use craft\base\FieldInterface;
 use craft\base\NestedElementInterface;
 use craft\behaviors\DraftBehavior;
 use craft\behaviors\RevisionBehavior;
+use craft\console\controllers\MigrateController;
+use craft\console\controllers\UpController;
+use craft\controllers\AppController;
 use craft\db\Connection;
 use craft\db\Query;
 use craft\db\QueryAbortedException;
@@ -1158,7 +1161,9 @@ class Elements extends Component
             ]));
         }
 
-        Db::delete(Table::ELEMENTS_BULKOPS, ['key' => $key], db: $this->bulkOpDb);
+        if (!$this->isMigrationRequest()) {
+            Db::delete(Table::ELEMENTS_BULKOPS, ['key' => $key], db: $this->bulkOpDb);
+        }
     }
 
     /**
@@ -1175,13 +1180,27 @@ class Elements extends Component
 
         $timestamp = Db::prepareDateForDb(DateTimeHelper::now());
 
-        foreach (array_keys($this->bulkKeys) as $key) {
-            Db::upsert(Table::ELEMENTS_BULKOPS, [
-                'elementId' => $element->id,
-                'key' => $key,
-                'timestamp' => $timestamp,
-            ], db: $this->bulkOpDb);
+        if (!$this->isMigrationRequest()) {
+            foreach (array_keys($this->bulkKeys) as $key) {
+                Db::upsert(Table::ELEMENTS_BULKOPS, [
+                    'elementId' => $element->id,
+                    'key' => $key,
+                    'timestamp' => $timestamp,
+                ], db: $this->bulkOpDb);
+            }
         }
+    }
+
+    private function isMigrationRequest(): bool
+    {
+        return (
+            Craft::$app->controller instanceof MigrateController ||
+            Craft::$app->controller instanceof UpController ||
+            (
+                Craft::$app->controller instanceof AppController &&
+                Craft::$app->controller->action?->id === 'update'
+            )
+        );
     }
 
     /**
